+++ src/BookRepository.php
@@ -82,20 +82,20 @@ class BookRepository
 
     public function deleteB($productsId = [])
     {
+        if ($productsId != null) {
+            foreach ($productsId as $productId) {
+                $sql = 'DELETE FROM book WHERE bookId = :productId';
 
-        foreach ($productsId as $productId) {
-            $sql = 'DELETE FROM book WHERE bookId = :productId';
+                // Prepare SQL
 
-            // Prepare SQL
+                $statement = $this->connection->prepare($sql);
 
-            $statement = $this->connection->prepare($sql);
+                // Bind parameters to statement variables
+                $statement->bindParam(':productId', $productId);
 
-            // Bind parameters to statement variables
-            $statement->bindParam(':productId', $productId);
-
-            // Execute statement
-            $noError = $statement->execute();
-
+                // Execute statement
+                $noError = $statement->execute();
+            }
         }
     }
